[chore] 패키징 수정
