+++ src/main/java/com/example/bolmalre/auth/service/CustomUserDetailsService.java
@@ -3,8 +3,8 @@ package com.example.bolmalre.auth.service;
 
 
 import com.example.bolmalre.auth.web.dto.CustomUserDetails;
-import com.example.bolmalre.member.Member;
-import com.example.bolmalre.member.MemberRepository;
+import com.example.bolmalre.member.domain.Member;
+import com.example.bolmalre.member.infrastructure.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
