+++ src/main/java/com/example/bolmalre/member/domain/Agreement.java
@@ -1,7 +1,8 @@
-package com.example.bolmalre.member;
+package com.example.bolmalre.member.domain;
 
 
 import com.example.bolmalre.common.domain.BaseEntity;
+import com.example.bolmalre.member.web.dto.MemberJoinDTO;
 import jakarta.persistence.*;
 import lombok.*;
 
@@ -33,7 +34,7 @@ public class Agreement extends BaseEntity {
     private Member member;
 
 
-    public static Agreement JoinDTOto(MemberJoinDTO.MemberJoinRequestDTO request,Member member){
+    public static Agreement JoinDTOto(MemberJoinDTO.MemberJoinRequestDTO request, Member member){
         return Agreement.builder()
                 .advAgreement(request.getAdvAgreement())
                 .financialAgreement(request.getFinancialAgreement())
