+++ src/main/java/com/example/bolmalre/member/domain/enums/Gender.java
@@ -1,0 +1,6 @@
+package com.example.bolmalre.member.domain.enums;
+
+public enum Gender {
+
+    MALE, FEMALE
+}
