+++ src/main/java/com/example/bolmalre/member/domain/enums/Role.java
@@ -1,0 +1,6 @@
+package com.example.bolmalre.member.domain.enums;
+
+public enum Role {
+
+    ROLE_USER,ROLE_ADMIN
+}
