+++ src/main/java/com/example/bolmalre/member/domain/enums/Status.java
@@ -1,0 +1,5 @@
+package com.example.bolmalre.member.domain.enums;
+
+public enum Status {
+    ACTIVE,INACTIVE
+}
