--- src/main/java/com/example/bolmalre/member/enums/Gender.java
@@ -1,6 +1,0 @@
-package com.example.bolmalre.member.enums;
-
-public enum Gender {
-
-    MALE, FEMALE
-}
