--- src/main/java/com/example/bolmalre/member/enums/Role.java
@@ -1,6 +1,0 @@
-package com.example.bolmalre.member.enums;
-
-public enum Role {
-
-    ROLE_USER,ROLE_ADMIN
-}
