--- src/main/java/com/example/bolmalre/member/enums/Status.java
@@ -1,5 +1,0 @@
-package com.example.bolmalre.member.enums;
-
-public enum Status {
-    ACTIVE,INACTIVE
-}
