+++ src/main/java/com/example/bolmalre/member/infrastructure/AgreementRepository.java
@@ -1,5 +1,6 @@
-package com.example.bolmalre.member;
+package com.example.bolmalre.member.infrastructure;
 
+import com.example.bolmalre.member.domain.Agreement;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AgreementRepository extends JpaRepository<Agreement, Long> {
