+++ src/main/java/com/example/bolmalre/member/web/port/MemberService.java
@@ -1,5 +1,6 @@
-package com.example.bolmalre.member;
+package com.example.bolmalre.member.web.port;
 
+import com.example.bolmalre.member.web.dto.MemberJoinDTO;
 import jakarta.validation.Valid;
 
 public interface MemberService {
